favor long for Total and Count like properties fix #466

+++ src/Nest/Domain/Facets/FacetItem.cs
@@ -2,9 +2,9 @@
 
 namespace Nest
 {
-    public abstract class FacetItem
-    {
-        [JsonProperty("count")]
-        public virtual int Count { get; internal set; }
-    }
+	public abstract class FacetItem
+	{
+		[JsonProperty("count")]
+		public virtual long Count { get; internal set; }
+	}
 }
