+++ src/Nest/Domain/Facets/FilterFacet.cs
@@ -2,9 +2,9 @@
 
 namespace Nest
 {
-    public class FilterFacet : Facet
-    {
-        [JsonProperty(PropertyName = "count")]
-        public int Count { get; internal set; }
-    }
+	public class FilterFacet : Facet
+	{
+		[JsonProperty(PropertyName = "count")]
+		public long Count { get; internal set; }
+	}
 }
