+++ src/Nest/Domain/Hit/HitsMetaData.cs
@@ -7,7 +7,7 @@ namespace Nest
     public class HitsMetaData<T> where T : class
     {
         [JsonProperty("total")]
-        public int Total { get; internal set; }
+        public long Total { get; internal set; }
 
         [JsonProperty("max_score")]
         public double MaxScore { get; internal set; }
