+++ src/Nest/Domain/Stats/IndexingStats.cs
@@ -10,21 +10,21 @@ namespace Nest
 	public class IndexingStats
 	{
 		[JsonProperty(PropertyName = "index_total")]
-		public int Total { get; set; }
+		public long Total { get; set; }
 		[JsonProperty(PropertyName = "index_time")]
 		public string Time { get; set; }
 		[JsonProperty(PropertyName = "index_time_in_millis")]
 		public double TimeInMilliseconds { get; set; }
 		[JsonProperty(PropertyName = "index_current")]
-		public int Current { get; set; }
+		public long Current { get; set; }
 		[JsonProperty(PropertyName = "delete_total")]
-		public int DeleteTotal { get; set; }
+		public long DeleteTotal { get; set; }
 		[JsonProperty(PropertyName = "delete_time")]
 		public string DeleteTime { get; set; }
 		[JsonProperty(PropertyName = "delete_time_in_millis")]
 		public double DeleteTimeInMilliseconds { get; set; }
 		[JsonProperty(PropertyName = "delete_current")]
-		public int DeleteCurrent { get; set; }
+		public long DeleteCurrent { get; set; }
 		[JsonProperty(PropertyName = "types")]
 		[JsonConverter(typeof(DictionaryKeysAreNotPropertyNamesJsonConverter))]
 		public Dictionary<string, TypeStats> Types { get; set; }
